quiz changes

+++ DODbot/main.py
@@ -12,6 +12,8 @@ def init_database():
  try:
     logger.info("🔑 Начало эксклюзивной инициализации БД")
     with db_manager.get_connection() as conn:
+        if conn is None:
+                raise RuntimeError("Соединение с БД не установлено")
         if os.path.exists("database.lock"):
             os.remove("database.lock")
         from newsletter import create_db
